move readmegetter to its own file
